Fix Index retro-compatibility
When reading an index created with an older version of the Plugin,
deserialization errors may raise, due to breaking changes introduced in
newer versions.

Such an index should be droppped (as unreadable) and a new one created
instead, using the new storage strategy.

To test this behavior, a new [Integration
test](src/it/indexRetroCompatibility/invoke.properties) have been introduced:
 * download a file with version 1.3.0 (creating an index based on the
 previous storage strategy)
 * download a file with the current version (build should not break)