Rename 'checkSignature' to 'alwaysVerifyChecksum'
Fixes #186, see the naming and documentation discussion there.

First, the plan was to just rename the one parameter and keep the one
with the old name around as deprecated. while doing so, I noticed lots
of
  - class and package names,
  - method names and parameters,
  - javadocs and comments,
referring to digest-based checksums like SHA-1 and MD5 as "signatures"
too, so I took care of globally renaming all of those too. That should
make it easier to work with the code base in the future.