Add support for wget unpackWhenChanged setting
Adds support for an unpackWhenChanged setting in the
WGet goal, which will only unpack a downloaded file
when it was not already present in the cache.

This is useful for executions which unpack downloads
into non-transient directories (i.e. not ./target)
which only need subsequent unpacking when the source
dependency archive changes.