Unpack directly from cached files when available (#284)
If downloaded files are archives to be unpacked and files
are cached, unpack the cached files directly instead of copying
them to the outputFile path first. This saves time and storage
and prevents strange behavior with the new unpackWhenChanged
option (cached files were being copied to outputFile and then
left there since the archives were not unpacked due to unchanged
downloads).